SQL validation for benchmarks
