+++ src/benchmark/tpch_benchmark.cpp
@@ -46,7 +46,7 @@ int main(int argc, char* argv[]) {
   cli_options.add_options()
     ("s,scale", "Database scale factor (1.0 ~ 1GB)", cxxopts::value<float>()->default_value("0.1"))
     ("q,queries", "Specify queries to run (comma-separated query ids, e.g. \"--queries 1,3,19\"), default is all", cxxopts::value<std::string>()) // NOLINT
-    ("use_prepared_statements", "Do not use prepared statements instead of random SQL strings", cxxopts::value<bool>()->default_value("true")); // NOLINT
+    ("use_prepared_statements", "Do not use prepared statements instead of random SQL strings", cxxopts::value<bool>()->default_value("false")); // NOLINT
   // clang-format on
 
   std::unique_ptr<opossum::BenchmarkConfig> config;
@@ -133,6 +133,8 @@ int main(int argc, char* argv[]) {
   config->out << "- ... done." << std::endl;
 
   auto context = opossum::BenchmarkRunner::create_context(*config);
+
+  Assert(!use_prepared_statements || !config->validate, "SQLite validation does not work with prepared statements");
 
   // Add TPCH-specific information
   context.emplace("scale_factor", scale_factor);
