+++ src/benchmarklib/benchmark_runner.hpp
@@ -21,6 +21,7 @@
 namespace opossum {
 
 class SQLPipeline;
+class SQLiteWrapper;
 
 class BenchmarkRunner {
  public:
@@ -83,6 +84,9 @@ class BenchmarkRunner {
   std::optional<PerformanceWarningDisabler> _performance_warning_disabler;
 
   Duration _total_run_duration{};
+
+  // If the query execution should be validated, this stores a pointer to the used SQLite instance
+  std::shared_ptr<SQLiteWrapper> _sqlite_wrapper;
 };
 
 }  // namespace opossum
