+++ src/lib/CMakeLists.txt
@@ -536,6 +536,8 @@ set(
     utils/aligned_size.hpp
     utils/assert.hpp
     utils/boost_default_memory_resource.cpp
+    utils/check_table_equal.cpp
+    utils/check_table_equal.hpp
     utils/ignore_unused_variable.hpp
     utils/copyable_atomic.hpp
     utils/enum_constant.hpp
@@ -561,6 +563,8 @@ set(
     utils/singleton.hpp
     utils/string_utils.cpp
     utils/string_utils.hpp
+    utils/sqlite_wrapper.cpp
+    utils/sqlite_wrapper.hpp
     utils/template_type.hpp
     utils/timer.cpp
     utils/timer.hpp
