+++ src/lib/utils/sqlite_wrapper.cpp
