+++ src/test/testing_assert.hpp
@@ -11,42 +11,12 @@
 #include "operators/abstract_operator.hpp"
 #include "scheduler/operator_task.hpp"
 #include "types.hpp"
+#include "utils/check_table_equal.hpp"
 
 namespace opossum {
 
 class AbstractLQPNode;
 class Table;
-
-/**
- * Indicates whether the comparison of two tables should happen order sensitive (Yes) or whether it should just be
- * checked whether both tables contain the same rows, independent of order.
- */
-enum class OrderSensitivity { Yes, No };
-
-/**
- * "Strict" enforces that both tables have precisely the same column types, "Lenient" allows float instead of double, double
- * instead of float, long instead of int, int instead of long
- */
-enum class TypeCmpMode { Strict, Lenient };
-
-/**
- * When comparing tables generated by Hyrise to those generated by, e.g. SQLite, minor differences are to be expected
- * (since sqlite uses double for arithmetics, Hyrise might use float) so for large numbers
- * FloatComparisonMode::RelativeDifference is better since it allows derivation independent of the absolute value.
- * When checking against manually generated tables, FloatComparisonMode::AbsoluteDifference is the better choice.
- */
-enum class FloatComparisonMode { RelativeDifference, AbsoluteDifference };
-
-// @{
-/**
- * Functions implementing the functionality of our custom macros defined at the bottom of this file.
- */
-
-bool check_table_equal(const std::shared_ptr<const Table>& opossum_table,
-                       const std::shared_ptr<const Table>& expected_table, OrderSensitivity order_sensitivity,
-                       TypeCmpMode type_cmp_mode, FloatComparisonMode float_comparison_mode);
-
-// @}
 
 bool check_lqp_tie(const std::shared_ptr<const AbstractLQPNode>& output, LQPInputSide input_side,
                    const std::shared_ptr<const AbstractLQPNode>& input);
