+++ src/test/tpc/tpch_test.cpp
@@ -60,7 +60,8 @@ TEST_P(TPCHTest, Test) {
   SCOPED_TRACE("TPC-H " + std::to_string(tpch_idx) + (use_jit ? " with JIT" : " without JIT") + " and " +
                (use_prepared_statements ? " with prepared statements" : " without prepared statements"));
 
-  auto query_generator = TPCHQueryGenerator{use_prepared_statements};
+  // The scale factor passed to the query generator will be ignored as we only use deterministic queries
+  auto query_generator = TPCHQueryGenerator{use_prepared_statements, 1.0f};
   if (use_prepared_statements) {
     // Run the preparation queries
     const auto& sql = query_generator.get_preparation_queries();
@@ -117,19 +118,19 @@ TEST_P(TPCHTest, Test) {
 }
 
 INSTANTIATE_TEST_CASE_P(TPCHTestNoJITNoPreparedStatements, TPCHTest,
-                        testing::Combine(testing::ValuesIn(TPCHQueryGenerator{false}.selected_queries()),
+                        testing::Combine(testing::ValuesIn(TPCHQueryGenerator{false, 1.0f}.selected_queries()),
                                          testing::ValuesIn({false}),
                                          testing::ValuesIn({false})), );  // NOLINT(whitespace/parens)
 
 INSTANTIATE_TEST_CASE_P(TPCHTestNoJITPreparedStatements, TPCHTest,
-                        testing::Combine(testing::ValuesIn(TPCHQueryGenerator{false}.selected_queries()),
+                        testing::Combine(testing::ValuesIn(TPCHQueryGenerator{false, 1.0f}.selected_queries()),
                                          testing::ValuesIn({false}),
                                          testing::ValuesIn({true})), );  // NOLINT(whitespace/parens)
 
 #if HYRISE_JIT_SUPPORT
 
 INSTANTIATE_TEST_CASE_P(TPCHTestJITPreparedStatements, TPCHTest,
-                        testing::Combine(testing::ValuesIn(TPCHQueryGenerator{false}.selected_queries()),
+                        testing::Combine(testing::ValuesIn(TPCHQueryGenerator{false, 1.0f}.selected_queries()),
                                          testing::ValuesIn({true}),
                                          testing::ValuesIn({true})), );  // NOLINT(whitespace/parens)
